(PC-35029)[PRO] refactor: move AddressFormValues to shared types
