+++ pro/src/pages/VenueSettings/types.ts
@@ -1,4 +1,4 @@
-import { AddressFormValues } from 'components/AddressManual/AddressManual'
+import { AddressFormValues } from 'commons/core/shared/types'
 
 export interface VenueSettingsFormValues extends AddressFormValues {
   bookingEmail: string
